Move map data into backtrace data proper.
The backtrace structure used to include a pointer to a backtrace_map_t
that represented the map data for a particular pc. This introduced a
race condition where the pointer could be discarded, but the backtrace
structure still contained a pointer to garbage memory. Now all of the map
information is right in the structure.

Bug: 19028453

Change-Id: I6021d1e6d265f246cafa6153e877fda52d04cc74
Signed-off-by: Chet Kener <Cl3Kener@gmail.com>